+++ c8y_api/model/inventory.py
@@ -23,11 +23,30 @@ class Inventory(CumulocityResource):
     def __init__(self, c8y):
         super().__init__(c8y, 'inventory/managedObjects')
 
-    def get(self, id) -> ManagedObject:  # noqa (id)
+    def get(
+            self,
+            id: str,  # noqa
+            with_children: bool = None,
+            with_children_count: bool = None,
+            skip_children_names: bool = None,
+            with_parents: bool = None,
+            with_latest_values: bool = None,
+            **kwargs) -> ManagedObject:
         """ Retrieve a specific managed object from the database.
 
         Args:
-            ID of the managed object
+            id (str): Cumulocity ID of the managed object
+            with_children (bool):  Whether children with ID and name should be
+                included with each returned object
+            with_children_count (bool): When set to true, the returned result
+                will contain the total number of children in the respective
+                child additions, assets and devices sub fragments.
+            skip_children_names (bool):  If true, returned references of child
+                devices won't contain their names.
+            with_parents (bool): Whether to include a device's parents.
+            with_latest_values (bool):  If true the platform includes the
+                fragment `c8y_LatestMeasurements, which contains the latest
+                measurement values reported by the device to the platform.
 
         Returns:
              A ManagedObject instance
@@ -35,7 +54,15 @@ class Inventory(CumulocityResource):
         Raises:
             KeyError:  if the ID is not defined within the database
         """
-        managed_object = ManagedObject.from_json(self._get_object(id))
+        managed_object = ManagedObject.from_json(self._get_object(
+            id,
+            with_children=with_children,
+            with_children_count=with_children_count,
+            skip_children_names=skip_children_names,
+            with_parents=with_parents,
+            with_latest_values=with_latest_values,
+            **kwargs)
+        )
         managed_object.c8y = self.c8y  # inject c8y connection into instance
         return managed_object
 
@@ -97,6 +124,66 @@ class Inventory(CumulocityResource):
             page_size=page_size,
             as_values=as_values,
             **kwargs))
+
+    def get_by(
+            self,
+            expression: str = None,
+            query: str = None,
+            ids: list[str | int] = None,
+            order_by: str = None,
+            type: str = None,
+            parent: str = None,
+            fragment: str = None,
+            fragments: list[str] = None,
+            name: str = None,
+            owner: str = None,
+            text: str = None,
+            only_roots: str = None,
+            with_children: bool = None,
+            with_children_count: bool = None,
+            skip_children_names: bool = None,
+            with_groups: bool = None,
+            with_parents: bool = None,
+            with_latest_values: bool = None,
+            as_values: str | tuple | list[str | tuple] = None,
+            **kwargs) -> ManagedObject:
+        """ Query the database for a specific managed object.
+
+        This function is a special version of the `select` function assuming a single
+        result being returned by the query.
+
+        Returns:
+            A ManagedObject instance
+
+        Raises:
+            ValueError:  if the query did not return any or more than one result.
+        """
+        result = list(self.select(
+            expression=expression,
+            query=query,
+            ids=ids,
+            order_by=order_by,
+            type=type,
+            parent=parent,
+            fragment=fragment,
+            fragments=fragments,
+            name=name,
+            owner=owner,
+            text=text,
+            only_roots=only_roots,
+            with_children=with_children,
+            with_children_count=with_children_count,
+            skip_children_names=skip_children_names,
+            with_groups=with_groups,
+            with_parents=with_parents,
+            with_latest_values=with_latest_values,
+            page_size=2,
+            as_values=as_values,
+            **kwargs))
+        if len(result) == 1:
+            return result[0]
+        raise ValueError("No matching object found." if not result
+                         else "Ambiguous query; multiple matching objects found.")
 
     def get_count(
             self,
@@ -455,11 +542,30 @@ class DeviceInventory(Inventory):
         """
         self.c8y.put('/devicecontrol/newDeviceRequests/' + str(id), {'status': 'ACCEPTED'})
 
-    def get(self, id: str) -> Device:  # noqa (id)
+    def get(
+            self,
+            id: str,  # noqa
+            with_children: bool = None,
+            with_children_count: bool = None,
+            skip_children_names: bool = None,
+            with_parents: bool = None,
+            with_latest_values: bool = None,
+            **kwargs) -> Device:
         """ Retrieve a specific device object.
 
         Args:
-            id (str): ID of the device object
+            id (str): Cumulocity ID of the device object
+            with_children (bool):  Whether children with ID and name should be
+                included with each returned object
+            with_children_count (bool): When set to true, the returned result
+                will contain the total number of children in the respective
+                child additions, assets and devices sub fragments.
+            skip_children_names (bool):  If true, returned references of child
+                devices won't contain their names.
+            with_parents (bool): Whether to include a device's parents.
+            with_latest_values (bool):  If true the platform includes the
+                fragment `c8y_LatestMeasurements, which contains the latest
+                measurement values reported by the device to the platform.
 
         Returns:
             A Device instance
@@ -467,7 +573,15 @@ class DeviceInventory(Inventory):
         Raises:
             KeyError:  if the ID is not defined within the database
         """
-        device = Device.from_json(self._get_object(id))
+        device = Device.from_json(self._get_object(
+            id,
+            with_children=with_children,
+            with_children_count=with_children_count,
+            skip_children_names=skip_children_names,
+            with_parents=with_parents,
+            with_latest_values=with_latest_values,
+            **kwargs)
+        )
         device.c8y = self.c8y
         return device