+++ tests/model/test_alarms.py
@@ -37,10 +37,10 @@ def isolate_call_url(fun, **kwargs):
 @pytest.mark.parametrize('params, expected, not_expected', [
     ({'expression': 'EX', 'type': 'T'}, ['?EX'], ['type']),
     ({'type': 'T', 'source': 'S', 'fragment': 'F'}, ['type=T', 'source=S', 'fragmentType=F'], []),
-    ({'status': 'ST', 'severity': 'SE', 'resolved': False}, ['status=ST', 'severity=SE', 'resolved=False'], []),
+    ({'status': 'ST', 'severity': 'SE', 'resolved': False}, ['status=ST', 'severity=SE', 'resolved=false'], []),
     ({'reverse': False, 'page_size': 8}, ['revert=false', 'pageSize=8'], ['reverse']),
     ({'with_source_assets': False, 'with_source_devices': True, 'source': '123'},
-     ['withSourceAssets=False', 'withSourceDevices=True'],
+     ['withSourceAssets=false', 'withSourceDevices=true'],
      ['_']),
     # data priorities
     ({'date_from': '2020-12-31', 'date_to': '2021-12-31'},